Added ComputerInterface.clp as well as a BaseClasses implementation
I realized that the majority of the work in implementing classes will have have
common components. Because of this, I have included a base classes file that is
loaded ahead of everything else. At this point it only has the has-flags
abstract class but others will be added as time goes on.